Restructure repo to feature-sliced design
