+++ src/features/validators-map/assets/continents.json
