+++ src/features/validators-map/utils/lineUtils.ts
