+++ src/pages/api/continents.ts
@@ -6,7 +6,7 @@ export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse<FeatureCollection>
 ) {
-  const continentsJson = await fs.readFile('./src/assets/continents.json', 'utf-8')
+  const continentsJson = await fs.readFile('./src/features/validators-map/assets/continents.json', 'utf-8')
   const continents = JSON.parse(continentsJson) as FeatureCollection
   res.status(200).json(continents)
 }
