+++ src/pages/index.tsx
@@ -2,7 +2,7 @@ import Head from 'next/head'
 import { Inter } from 'next/font/google'
 import styles from '@/styles/Home.module.css'
 
-import ValidatorsMap from '@/components/ValidatorsMap'
+import ValidatorsMap from '@/widgets/ValidatorsMap'
 
 const inter = Inter({ subsets: ['latin'] })
